Refactor Akka.TestKit.Xunit3 to Akka.TestKit.Xunit
