+++ src/contrib/testkits/Akka.TestKit.Xunit.Tests/Internals/AkkaEqualExceptionSpec.cs
@@ -5,10 +5,10 @@
 // </copyright>
 //-----------------------------------------------------------------------
 
-using Akka.TestKit.Xunit3.Internals;
+using Akka.TestKit.Xunit.Internals;
 using Xunit;
 
-namespace Akka.TestKit.Xunit2.Tests.Internals;
+namespace Akka.TestKit.Xunit.Tests.Internals;
 
 public static class AkkaEqualExceptionSpec
 {
