+++ src/contrib/testkits/Akka.TestKit.Xunit.Tests/XunitAssertionsSpec.cs
@@ -5,10 +5,11 @@
 // </copyright>
 //-----------------------------------------------------------------------
 
+using Akka.TestKit.Xunit;
 using Xunit;
 using Xunit.Sdk;
 
-namespace Akka.TestKit.Xunit3.Tests;
+namespace Akka.TestKit.Xunit.Tests;
 
 public class XunitAssertionsSpec
 {
