+++ src/contrib/testkits/Akka.TestKit.Xunit/Attributes/LocalFactAttribute.cs
@@ -5,10 +5,11 @@
 // </copyright>
 //-----------------------------------------------------------------------
 
+using System;
 using Xunit;
 using Xunit.v3;
 
-namespace Akka.TestKit.Xunit3.Attributes;
+namespace Akka.TestKit.Xunit.Attributes;
 
 /// <summary>
 /// <para>
