+++ src/contrib/testkits/Akka.TestKit.Xunit/TestKit.cs
@@ -11,159 +11,211 @@ using Akka.Actor.Setup;
 using Akka.Configuration;
 using Akka.Event;
 using Akka.TestKit.Xunit.Internals;
-using Xunit.Abstractions;
+using Xunit;
 
-namespace Akka.TestKit.Xunit
+namespace Akka.TestKit.Xunit;
+
+/// <summary>
+/// This class represents an Akka.NET TestKit that uses <a href="https://xunit.github.io/">xUnit</a>
+/// as its testing framework.
+/// </summary>
+public class TestKit : TestKitBase, IDisposable
 {
-    /// <summary>
-    /// This class represents an Akka.NET TestKit that uses <a href="https://xunit.github.io/">xUnit</a>
-    /// as its testing framework.
-    /// </summary>
-    public class TestKit : TestKitBase , IDisposable
+    private class PrefixedOutput : ITestOutputHelper
     {
-        private bool _isDisposed; //Automatically initialized to false;
+        private static readonly string Newline = Environment.NewLine;
+        private readonly ITestOutputHelper _output;
+        private readonly string _prefix;
+        private bool _newLine;
 
-        /// <summary>
-        /// The provider used to write test output.
-        /// </summary>
-        protected readonly ITestOutputHelper Output;
-
-        /// <summary>
-        /// <para>
-        /// Initializes a new instance of the <see cref="TestKit"/> class.
-        /// </para>
-        /// <para>
-        /// If no <paramref name="system"/> is passed in, a new system with
-        /// <see cref="DefaultConfig"/> will be created.
-        /// </para>
-        /// </summary>
-        /// <param name="system">The actor system to use for testing. The default value is <see langword="null"/>.</param>
-        /// <param name="output">The provider used to write test output. The default value is <see langword="null"/>.</param>
-        public TestKit(ActorSystem system = null, ITestOutputHelper output = null)
-            : base(Assertions, system)
+        public PrefixedOutput(ITestOutputHelper output, string prefix)
         {
-            Output = output;
-            InitializeLogger(Sys);
+            _output = output;
+            _prefix = prefix;
         }
 
-        /// <summary>
-        /// Initializes a new instance of the <see cref="TestKit"/> class.
-        /// </summary>
-        /// <param name="config">The <see cref="Setup"/> to use for configuring the ActorSystem.</param>
-        /// <param name="actorSystemName">The name of the system. The default name is "test".</param>
-        /// <param name="output">The provider used to write test output. The default value is <see langword="null"/>.</param>
-        public TestKit(ActorSystemSetup config, string actorSystemName = null, ITestOutputHelper output = null)
-            : base(Assertions, config, actorSystemName)
+        public void Write(string message)
         {
-            Output = output;
-            InitializeLogger(Sys);
+            if(_newLine)
+                message = _prefix + Newline;
+                
+            var parts = message.Split([Newline], StringSplitOptions.None);
+            _output.Write(string.Join($"{Newline}{_prefix}", parts));
+            _newLine = message.EndsWith(Newline);
         }
 
-        /// <summary>
-        /// Initializes a new instance of the <see cref="TestKit"/> class.
-        /// </summary>
-        /// <param name="config">The configuration to use for the system.</param>
-        /// <param name="actorSystemName">The name of the system. The default name is "test".</param>
-        /// <param name="output">The provider used to write test output. The default value is <see langword="null"/>.</param>
-        public TestKit(Config config, string actorSystemName = null, ITestOutputHelper output = null)
-            : base(Assertions, config, actorSystemName)
+        public void Write(string format, params object[] args)
         {
-            Output = output;
-            InitializeLogger(Sys);
+            if(_newLine)
+                format = _prefix + format;
+            var parts = format.Split([Newline], StringSplitOptions.None);
+            format = string.Join($"{Newline}{_prefix}", parts);
+                
+            _output.Write(format, args);
+            _newLine = format.EndsWith(Newline);
         }
 
-        /// <summary>
-        /// Initializes a new instance of the <see cref="TestKit"/> class.
-        /// </summary>
-        /// <param name="config">The configuration to use for the system.</param>
-        /// <param name="output">The provider used to write test output. The default value is <see langword="null"/>.</param>
-        public TestKit(string config, ITestOutputHelper output = null)
-            : base(Assertions, ConfigurationFactory.ParseString(config))
+        public void WriteLine(string message)
         {
-            Output = output;
-            InitializeLogger(Sys);
+            if (_newLine)
+                message = _prefix + message;
+            _newLine = true;
+            _output.WriteLine(message);
         }
 
-        /// <summary>
-        /// A configuration that has just the default log settings enabled. The default settings can be found in
-        /// <a href="https://github.com/akkadotnet/akka.net/blob/master/src/core/Akka.TestKit/Internal/Reference.conf">Akka.TestKit.Internal.Reference.conf</a>.
-        /// </summary>
-        public new static Config DefaultConfig => TestKitBase.DefaultConfig;
+        public void WriteLine(string format, params object[] args)
+        {
+            if (_newLine)
+                format = _prefix + format;
+            _newLine = true;
+            _output.WriteLine(format, args);
+        }
 
-        /// <summary>
-        /// A configuration that has all log settings enabled
-        /// </summary>
-        public new static Config FullDebugConfig => TestKitBase.FullDebugConfig;
+        public string Output => _output.Output;
+    }
 
-        /// <summary>
-        /// Commonly used assertions used throughout the testkit.
-        /// </summary>
-        protected static XunitAssertions Assertions { get; } = new();
+    /// <summary>
+    /// The provider used to write test output.
+    /// </summary>
+    protected readonly ITestOutputHelper? Output;
 
-        /// <summary>
-        /// This method is called when a test ends.
-        /// 
-        /// <remarks>
-        /// If you override this, then make sure you either call <c>base.AfterTest()</c> or
-        /// <c>TestKitBase.Shutdown</c>
-        /// to shut down the system. Otherwise a memory leak will occur.
-        /// </remarks>
-        /// </summary>
-        protected virtual void AfterAll()
+    private bool _disposed;
+    private bool _disposing;
+        
+    /// <summary>
+    /// <para>
+    /// Initializes a new instance of the <see cref="TestKit"/> class.
+    /// </para>
+    /// <para>
+    /// If no <paramref name="system"/> is passed in, a new system with
+    /// <see cref="DefaultConfig"/> will be created.
+    /// </para>
+    /// </summary>
+    /// <param name="system">The actor system to use for testing. The default value is <see langword="null"/>.</param>
+    /// <param name="output">The provider used to write test output. The default value is <see langword="null"/>.</param>
+    public TestKit(ActorSystem? system = null, ITestOutputHelper? output = null)
+        : base(Assertions, system)
+    {
+        Output = output;
+        InitializeLogger(Sys);
+    }
+
+    /// <summary>
+    /// Initializes a new instance of the <see cref="TestKit"/> class.
+    /// </summary>
+    /// <param name="config">The <see cref="Setup"/> to use for configuring the ActorSystem.</param>
+    /// <param name="actorSystemName">The name of the system. The default name is "test".</param>
+    /// <param name="output">The provider used to write test output. The default value is <see langword="null"/>.</param>
+    public TestKit(ActorSystemSetup config, string? actorSystemName = null, ITestOutputHelper? output = null)
+        : base(Assertions, config, actorSystemName)
+    {
+        Output = output;
+        InitializeLogger(Sys);
+    }
+
+    /// <summary>
+    /// Initializes a new instance of the <see cref="TestKit"/> class.
+    /// </summary>
+    /// <param name="config">The configuration to use for the system.</param>
+    /// <param name="actorSystemName">The name of the system. The default name is "test".</param>
+    /// <param name="output">The provider used to write test output. The default value is <see langword="null"/>.</param>
+    public TestKit(Config config, string? actorSystemName = null, ITestOutputHelper? output = null)
+        : base(Assertions, config, actorSystemName)
+    {
+        Output = output;
+        InitializeLogger(Sys);
+    }
+
+    /// <summary>
+    /// Initializes a new instance of the <see cref="TestKit"/> class.
+    /// </summary>
+    /// <param name="config">The configuration to use for the system.</param>
+    /// <param name="output">The provider used to write test output. The default value is <see langword="null"/>.</param>
+    public TestKit(string config, ITestOutputHelper? output = null)
+        : base(Assertions, ConfigurationFactory.ParseString(config))
+    {
+        Output = output;
+        InitializeLogger(Sys);
+    }
+
+    /// <summary>
+    /// A configuration that has just the default log settings enabled. The default settings can be found in
+    /// <a href="https://github.com/akkadotnet/akka.net/blob/master/src/core/Akka.TestKit/Internal/Reference.conf">Akka.TestKit.Internal.Reference.conf</a>.
+    /// </summary>
+    public new static Config DefaultConfig => TestKitBase.DefaultConfig;
+
+    /// <summary>
+    /// A configuration that has all log settings enabled
+    /// </summary>
+    public new static Config FullDebugConfig => TestKitBase.FullDebugConfig;
+
+    /// <summary>
+    /// Commonly used assertions used throughout the testkit.
+    /// </summary>
+    protected static XunitAssertions Assertions { get; } = new();
+
+    /// <summary>
+    /// This method is called when a test ends.
+    /// </summary>
+    protected virtual void AfterAll()
+    {
+    }
+
+    /// <summary>
+    /// Initializes a new <see cref="TestOutputLogger"/> used to log messages.
+    /// </summary>
+    /// <param name="system">The actor system used to attach the logger</param>
+    protected void InitializeLogger(ActorSystem system)
+    {
+        if (Output == null) 
+            return;
+            
+        var extSystem = (ExtendedActorSystem)system;
+        var logger = extSystem.SystemActorOf(Props.Create(() => new TestOutputLogger(Output)), "log-test");
+        logger.Ask<LoggerInitialized>(new InitializeLogger(system.EventStream), TestKitSettings.TestKitStartupTimeout)
+            .ConfigureAwait(false).GetAwaiter().GetResult();
+    }
+
+    protected void InitializeLogger(ActorSystem system, string prefix)
+    {
+        if (Output == null) 
+            return;
+            
+        var extSystem = (ExtendedActorSystem)system;
+        var logger = extSystem.SystemActorOf(Props.Create(() => new TestOutputLogger(
+            string.IsNullOrEmpty(prefix) ? Output : new PrefixedOutput(Output, prefix))), "log-test");
+        logger.Ask<LoggerInitialized>(new InitializeLogger(system.EventStream), TestKitSettings.TestKitStartupTimeout)
+            .ConfigureAwait(false).GetAwaiter().GetResult();
+    }
+
+    /// <summary>
+    /// Performs application-defined tasks associated with freeing, releasing, or resetting unmanaged resources.
+    /// </summary>
+    /// <param name="disposing">
+    /// if set to <c>true</c> the method has been called directly or indirectly by a  user's code.
+    /// Managed and unmanaged resources will be disposed.<br /> if set to <c>false</c> the method
+    /// has been called by the runtime from inside the finalizer and only unmanaged resources can
+    ///  be disposed.
+    /// </param>
+    protected virtual void Dispose(bool disposing)
+    {
+        if (_disposing || _disposed)
+            return;
+            
+        _disposing = true;
+        try
+        {
+            AfterAll();
+        }
+        finally
         {
             Shutdown();
+            _disposed = true;
         }
+    }
 
-        /// <summary>
-        /// Initializes a new <see cref="TestOutputLogger"/> used to log messages.
-        /// </summary>
-        /// <param name="system">The actor system used to attach the logger</param>
-        protected void InitializeLogger(ActorSystem system)
-        {
-            if (Output != null)
-            {
-                var extSystem = (ExtendedActorSystem)system;
-                var logger = extSystem.SystemActorOf(Props.Create(() => new TestOutputLogger(Output)), "log-test");
-                logger.Tell(new InitializeLogger(system.EventStream));
-            }
-        }
-
-        
-        public void Dispose()
-        {
-            Dispose(true);
-            //Take this object off the finalization queue and prevent finalization code for this object
-            //from executing a second time.
-            GC.SuppressFinalize(this);
-        }
-
-        /// <summary>
-        /// Performs application-defined tasks associated with freeing, releasing, or resetting unmanaged resources.
-        /// </summary>
-        /// <param name="disposing">
-        /// if set to <c>true</c> the method has been called directly or indirectly by a  user's code.
-        /// Managed and unmanaged resources will be disposed.<br /> if set to <c>false</c> the method
-        /// has been called by the runtime from inside the finalizer and only unmanaged resources can
-        ///  be disposed.
-        /// </param>
-        protected virtual void Dispose(bool disposing)
-        {
-            // If disposing equals false, the method has been called by the
-            // runtime from inside the finalizer and you should not reference
-            // other objects. Only unmanaged resources can be disposed.
-
-            try
-            {
-                //Make sure Dispose does not get called more than once, by checking the disposed field
-                if(!_isDisposed && disposing)
-                {
-                    AfterAll();
-                }
-                _isDisposed = true;
-            }
-            finally
-            {
-            }
-        }
+    public void Dispose()
+    {
+        Dispose(true);
     }
 }
