--- src/contrib/testkits/Akka.TestKit.Xunit3/Akka.TestKit.Xunit3.csproj
@@ -1,20 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFrameworks>$(NetStandardLibVersion);net8.0</TargetFrameworks>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\..\..\core\Akka.TestKit\Akka.TestKit.csproj" />
-    <PackageReference Include="xunit.v3.extensibility.core" Version="$(Xunit3Version)" />
-    <PackageReference Include="xunit.v3.assert" Version="$(Xunit3Version)" />
-    <!-- <PackageReference Include="xunit.v3" Version="$(Xunit3Version)" /> -->
-  </ItemGroup>
-
-  <PropertyGroup>
-    <!-- WORKAROUND: for some reason starting at Akka.NET 1.3.2 this package was determined as "unpackable" by default via DOTNET CLI -->
-    <IsPackable>true</IsPackable>
-  </PropertyGroup>
-</Project>
